feat(plc): Phase 2 part A — extract NotesBody/TodosBody + fullscreen tile expansion
Extract reusable bodies from `PlcNotesTab` and `PlcTodosTab` into a new
`components/plc/bodies/` directory. The legacy tabs become 3-line
wrappers around the body so a single editor implementation feeds both
the v1 tab routing and the v2 grid's inline rendering.

Wire fullscreen tile expansion in the v2 grid:
- `PlcGridTile` accepts an optional `onExpand(kind)` callback and
  surfaces a top-right expand button on hover/focus when provided.
- `PlcGridLayout` gates the expand handler to tile kinds that have a
  real fullscreen body (notes, todos, quizLibrary, assignments,
  completedAssignments) — preview-only tiles (plcInfo, sharedSheet)
  don't get an expand affordance.
- `PlcDashboard` owns `expandedTile` state. When set, the dashboard
  body shows a fullscreen overlay with the tile's body + a back button.
  Escape collapses the overlay before falling back to the dashboard's
  own close behavior.

Quiz library and assignments bodies are left in the existing tab files
for now (they're more complex — multi-pane editors, sub-tabs); their
fullscreen path falls through to the existing tab components. Phase 2b
will extract them properly.

Bodies fully replace tab JSX (no behavioral changes); 2319 tests pass.
